% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multised.engine-package.R
\docType{package}
\name{multised.engine-package}
\alias{multised.engine}
\alias{multised.engine-package}
\title{multised.engine: Build and Analyse the MULTISED Marine Sediment Database}
\description{
Assembles a database of marine-sediment trace-element measurements from five external sources (Mareano, Vannmiljoe, ICES-DOME, MUDAB and 4Demon), harmonises them into a common schema, and progressively quality-controls the result through five database generations: pilot, slim, clean, merged and refined. Also provides the analyses run on the finished databases, covering grain size, iron and aluminium normalisation, organic carbon, spatial and temporal coverage, outlier flagging, and pristine background classification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seafood-hazards/multised-engine}
  \item Report bugs at \url{https://github.com/seafood-hazards/multised-engine/issues}
}

}
\author{
\strong{Maintainer}: Takaya Saito \email{takaya.saito@outlook.com} [copyright holder]

Authors:
\itemize{
  \item Takaya Saito \email{takaya.saito@outlook.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Institute of Marine Research [funder]
}

}
\keyword{internal}
